Legal_agent/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo_search
fastapi
uvicorn
groq
co